+++ hosts/armored-armadillo/default.nix
@@ -14,6 +14,12 @@
   ghostship.host.roles = {
     develop = true;
     wsl = true;
+    t3worker = true;
+  };
+
+  ghostship.t3Worker.directTunnel = {
+    enable = true;
+    hostname = "armored-armadillo-t3.ghostship.io";
   };
 
   wsl.defaultUser = "nixos";