+++ modules/agent-worker/default.nix
@@ -1,0 +1,9 @@
+{ ... }:
+
+{
+  imports = [
+    ./options.nix
+    ./tooling.nix
+    ./t3code-worker.nix
+  ];
+}